requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
black==21.9b0
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.8
django-rest-framework==0.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
flake8==4.0.1
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
kombu==5.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.0
pathspec==0.9.0
platformdirs==2.4.0
prompt-toolkit==3.0.21
psycopg2==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.1
python-dateutil==1.5
pytz==2021.3
regex==2021.10.23
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
